// server/utils/types.ts
export type JsonMeasureUnit = 'px' | 'mm' | 'cm' | 'in'

export type JsonUrlTarget = '_blank' | '_self' | '_parent' | '_top'

export const JsonLayerTypeArr = [
  'embed',
  'image',
  'shape',
  'svg',
  'video',
  'audio',
  'text',
  'button',
  'youtube',
  'clipart',
  'countdown',
  'rain',
  'snow',
  'fog',
] as const

export const NotAcceptedLayer = ['menu']

export type JsonLayerType = (typeof JsonLayerTypeArr)[number]

export type JsonDesignBackground = JsonBackground & {
  borderColor?: string
  useBorder: boolean
}

export type EditorLayerFeed = Record<string, string | null>
export type AdStudioLayerFeedPropertyBind = {
  hash: string
  column: string
  index?: string
} | null
export type AdStudioLayerFeed = Record<
  string,
  AdStudioLayerFeedPropertyBind | AdStudioLayerFeedPropertyBind[]
>

export type JsonPropertiesFeed = EditorLayerFeed | AdStudioLayerFeed
// ...
export type EditorDesignFeed = {
  hash?: string
  row?: number
  name?: string
  url?: string
  background?: string
}

export type AdStudioDesignFeed = {
  hashes?: { hash: string; row: number }[]
  properties?: {
    [key: string]: AdStudioDesignFeedProperty
  }
}

export type AdStudioDesignFeedProperty = { hash: string; column: string }

export type JsonDesignFeed = EditorDesignFeed | AdStudioDesignFeed

export function isEditorDesignFeed(
  variable: JsonDesignFeed
): variable is EditorDesignFeed {
  return typeof variable === 'object' && variable !== null && 'hash' in variable
}

export function isAdStudioDesignFeed(
  variable: JsonDesignFeed
): variable is AdStudioDesignFeed {
  return (
    typeof variable === 'object' && variable !== null && 'hashes' in variable
  )
}

export type JsonDesignProperties = {
  sortIndex?: number
  name: string
  width: number
  height: number
  loop: boolean
  loopCount?: number
  imageQuality?: number
  bannerSize?: string
  customSize?: boolean
  presetSize?: boolean
  backgroundColor?: JsonDesignBackground
  lastId: number
  version: string
  showGuidelines: boolean
  measureUnit: JsonMeasureUnit
  sizeVariation?: string | null
  feed?: JsonDesignFeed
  fallbackImagePosition?: number
  fallbackImageSlideIndex?: number
  bannerUrl: string
  useAsClickTag: boolean
}

export type JsonRGradPos =
  | 'center center'
  | 'center right'
  | 'center left'
  | 'bottom left'
  | 'bottom center'
  | 'bottom right'
  | 'top left'
  | 'top center'
  | 'top right'

export type JsonAlign =
  | 'bottom'
  | 'center'
  | 'left'
  | 'middle'
  | 'right'
  | 'top'

export type JsonScaleMode =
  | 'stretch'
  | 'aspect'
  | 'mask'
  | 'tile'
  | 'crop'
  | 'userCrop'

export type JsonBackgroundType = 'solid' | 'rgrad' | 'lgrad' | 'image' | 'none'

export type JsonBackgroundSolid = {
  type: 'solid'
  scolor: string
}

export type JsonBackgroundSolidText = JsonBackgroundSolid & {
  feed?: AdStudioLayerFeedPropertyBind
}

export type JsonBackgroundLinearGradient = {
  type: 'lgrad'
  gradColors: JsonGradColor[]
  backgroundRotation: number
}

export type JsonBackgroundRadialGradient = {
  type: 'rgrad'
  gradColors: JsonGradColor[]
  rgradPos: JsonRGradPos
}

export type JsonBackgroundImage = {
  type: 'image'
  originalWidth: number
  originalHeight: number
  url: string
  backgroundRotation?: number
  scaleMode: JsonScaleMode
  contentScale?: number
  horizontalAlign: JsonAlign
  verticalAlign: JsonAlign
  contentOffsetX?: number
  contentOffsetY?: number
  hqUrl?: string
  source?: string
  hash?: string
  licensing?: boolean
}

export type JsonBackgroundNone = {
  type: 'none'
}

export interface JsonBackgroundBorder {
  useBorder?: boolean
  borderColor?: string
}

export type JsonBackground =
  | JsonBackgroundNone
  | JsonBackgroundSolid
  | JsonBackgroundLinearGradient
  | JsonBackgroundRadialGradient
  | JsonBackgroundImage

export type JsonBackgroundWithBorder = JsonBackgroundBorder & JsonBackground

export type JsonTextGradient =
  | JsonBackgroundNone
  | JsonBackgroundSolidText
  | JsonBackgroundLinearGradient
  | JsonBackgroundRadialGradient
  | JsonBackgroundImage
// ...
export type JsonTransition = {
  type: JsonTransitionType
  direction: JsonTransitionDirection
  duration: number
  delay: number
  tween: JsonTweenType
  ease: JsonEaseType
  wordsDuration?: number
  wordsAppearOrder?: 'l2r' | 'r2l'
  slidePosX: number
  slidePosY: number
  alphaOffset: number
  slideOffset: number
  blurAmount: number
  zoom: number
  preset: string
  color?: string
  shadowLength?: number
  advancedProperties?: AdvancedProperties
}

export type JsonTransitionGradColor = {
  gradColors: JsonGradColor[]
}

export type JsonTransitionMid = {
  type: JsonTransitionType
  direction: 'vertical' | 'horizontal' | 'top'
  variation?: string | null
  duration: number
  delay: number
  motionTime: number
  ease: JsonEaseType
  tween: JsonTweenType
  preset: string
  startEnd: string
  advancedProperties?: AdvancedProperties
  gradient?: JsonTransitionGradColor
}
// ...
export type JsonElementProperties =
  | JsonTextProperties
  | JsonButtonProperties
  | JsonImageProperties
  | JsonSvgProperties
  | JsonAudioProperties
  | JsonVideoProperties
  | JsonShapeProperties
  | JsonEmbedProperties
  | JsonYoutubeProperties
  | JsonClipartProperties
  | JsonRainProperties
  | JsonSnowProperties
  | JsonFogProperties
  | JsonCountdownProperties

export type JsonShadow = {
  useShadow: boolean
  hShadow: number
  vShadow: number
  blur: number
  spread: number
  color: string
}

export type JsonBlur = {
  useBlur: boolean
  pixels: number
}

export const JsonFlipTypeArr = [
  'both',
  'horizontal',
  'none',
  'vertical',
] as const

export type JsonFlipType = (typeof JsonFlipTypeArr)[number]

export type JsonTextAlignment = 'left' | 'right' | 'center' | 'justify'

export const JsonTextVerticalAlignment = {
  top: 'top',
  center: 'center',
  bottom: 'bottom',
} as const
export type JsonTextVerticalAlignmentType =
  (typeof JsonTextVerticalAlignment)[keyof typeof JsonTextVerticalAlignment]
// ...
export const JsonTextContentHeight = {
  auto: 'auto',
  fixed: 'fixed',
} as const
export type JsonTextContentHeightType =
  (typeof JsonTextContentHeight)[keyof typeof JsonTextContentHeight]
// ...
export type JsonAction =
  | JsonActionGotoURL
  | JsonActionGotoSlide
  | JsonActionNone
  | JsonActionVideoEnd

export function isJsonActionGotoSlide(
  action: JsonAction
): action is JsonActionGotoSlide {
  return action.type === 'gotoSlide'
}

export function isJsonActionGotoURL(
  action: JsonAction
): action is JsonActionGotoURL {
  return action.type === 'gotoURL'
}

export function isJsonActionNone(action: JsonAction): action is JsonActionNone {
  return action.type === 'none'
}

export type JsonBaseElementProperties = {
  id: number
  bannersetElementId?: number | null
  elementCategory?: string | null
  layerName: string
  dropShadow?: JsonShadow
  blur?: JsonBlur
  opacity: number
  rotation: number
  flip?: JsonFlipType
  x: number
  y: number
  width: number
  height: number
  buildIn?: JsonTransition
  buildMid?: JsonTransitionMid
  buildOut?: JsonTransition
  actions?: JsonAction[]
  locked?: boolean
  lockedProperties?: string[]
  visible: boolean
  group?: number | null
  feed?: JsonPropertiesFeed
}

export type JsonTextProperties = JsonBaseElementProperties & {
  text?: string
  alignment: JsonTextAlignment
  verticalAlign?: JsonTextVerticalAlignmentType
  contentHeightType: JsonTextContentHeightType
  fontSize: number
  lineHeight: number
  letterSpacing: number
  config: JsonTextSlateConfig // TODO:
  textShadow?: JsonShadow
  initialFontSize?: number
  textDirection?: string | null
  scale?: number
  blendMode?: JsonBlendMode
  outline?: JsonTextOutline
}

export type JsonTextOutline = {
  useOutline?: boolean
  color?: string
  weight?: number
}
// ...
export type JsonFontSettings = {
  fontFamily: string
  fontType: FontType
  fontStyle: FontStyle
  fontPrefix: string | null
  fontWeight: number
  fontFaceUrl: string | null
  fontUrl: string | null
}

export type JsonTextSlateConfigChildren = {
  color: string
  textScript: 'none' | 'superscript' | 'subscript'
  textDecoration: TextDecorationEnum
  textTransform: TextTransform
  fontSettings: JsonFontSettings
  typography: 'heading' | 'subheading' | 'body'
  gradient?: JsonTextGradient
  fontSizePercent: number
  fontSize?: number
  text: string
  feed?: AdStudioLayerFeedPropertyBind
  feedText?: AdStudioLayerFeedPropertyBind
}

export enum TextContentBoundaryType {
  OVERFLOW = 'overflow',
  SHRINK = 'shrink',
  ELLIPSIS = 'ellipsis',
}

export type Boundary = {
  type?: JsonTextContentBoundaryType
  minFontSize?: number
}

export type JsonTextSlateConfig = {
  type: 'slate'
  nodes: JsonTextSlateConfigNode[]
  selection?: unknown
  keepSelection?: boolean
  boundary?: Boundary
}
// ...
export type JsonCountdownConfig = {
  style: JsonCountdownConfigStyle
  children: { style: JsonCountdownConfigStyle }[]
}

export type JsonCountdownProperties = JsonBaseElementProperties & {
  countToDate: boolean
  countValue: string | number
  countTimeZone: number
  fontSize: number
  letterSpacing: number
  lineHeight: number
  scale: number
  textShadow: JsonShadow
  alignment: JsonTextAlignment
  config: JsonCountdownConfig[]
}

export type JsonCountdown = {
  type: 'layer'
  layerType: 'countdown'
  properties: JsonCountdownProperties
}

export type JsonRainProperties = JsonBaseElementProperties & {
  size: number
  speed: number
  density: number
  opacity: number
  particleSize: number
  splashEffect: boolean
  splashEffectParticleSize: number
  splashEffectSize: number
}

export type JsonRain = {
  type: 'layer'
  layerType: 'rain'
  properties: JsonRainProperties
}

export type JsonSnowProperties = JsonBaseElementProperties & {
  size: number
  speed: number
  density: number
  opacity: number
  particleSize: number
}

export type JsonSnow = {
  type: 'layer'
  layerType: 'snow'
  properties: JsonSnowProperties
}

export type JsonFogProperties = JsonBaseElementProperties & {
  size: number
  speed: number
  density: number
  opacity: number
}

export type JsonFog = {
  type: 'layer'
  layerType: 'fog'
  properties: JsonFogProperties
}

export type JsonBorder = {
  color: string
  weight: number
  radius: number
}

export type FontType = 'google' | 'custom' | 'all' | 'bannersnack' | 'creatopy'
export type FontStyle = 'normal' | 'italic' | 'regular' | 'slanted' | 'oblique'
export type TextTransform = 'none' | 'uppercase' | 'lowercase' | 'capitalize'

export type JsonButtonLabelStyle = JsonFontSettings & {
  fontDesign?: string
  textTransform?: TextTransform
  fontSize?: number
  color?: string
  letterSpacing?: number
  dropShadow?: JsonShadow
  initialFontSize?: number
  textDirection?: string
  fontPrefix?: string | null
}

export type JsonButtonProperties = JsonBaseElementProperties & {
  buttonLabel: string
  html: string
  backgroundColor?: JsonBackground
  backgroundOverColor?: string
  labelShadow?: JsonShadow
  labelStyle: JsonButtonLabelStyle
  labelOffsetX: number
  labelOffsetY: number
  border?: JsonBorder
  hoverState?: {
    backgroundColor?: JsonBackground
    border?: JsonBorder
    dropShadow?: JsonShadow
    labelShadow?: JsonShadow
    labelStyle?: JsonButtonLabelStyle
  }
}

export type JsonButton = {
  type: 'layer'
  layerType: 'button'
  properties: JsonButtonProperties
}

export type JsonAdjustColor = {
  useAdjustColor: boolean
  brightness: number
  contrast: number
  saturate: number
  hue: number
}

export type JsonCropData = {
  x: number
  y: number
  width: number
  height: number
}

export type JsonImageProperties = JsonBaseElementProperties & {
  url: string
  scaleMode: JsonScaleMode
  horizontalAlign: JsonAlign
  verticalAlign: JsonAlign
  adjustColor: JsonAdjustColor
  contentScale: number
  contentOffsetX: number
  contentOffsetY: number
  originalWidth: number
  originalHeight: number
  hqUrl?: string
  cropData?: JsonCropData
  source?: string
  hash?: string
  licensing?: boolean
  blendMode?: JsonBlendMode
  mediaFolderId?: number
  originalName?: string
  maskImage?: string
}

export type JsonImage = {
  type: 'layer'
  layerType: 'image'
  properties: JsonImageProperties
}

export type JsonLineData = {
  iteration: number
  width: number
  height: number
  left: number
  leftMargin: number
  rightMargin: number
  leftWidth: number
  rightWidth: number
}

export type JsonSvgProperties = JsonBaseElementProperties & {
  url: string
  colorGroups: Record<string, string>
  originalWidth: number
  originalHeight: number
  resourceKey: string
  lineData?: JsonLineData // TODO
}

export type JsonSvg = {
  type: 'layer'
  layerType: 'svg'
  properties: JsonSvgProperties
}

export type JsonBlendMode =
  | 'color'
  | 'color-burn'
  | 'color-dodge'
  | 'darken'
  | 'difference'
  | 'exclusion'
  | 'hard-light'
  | 'hue'
  | 'lighten'
  | 'luminosity'
  | 'multiply'
  | 'normal'
  | 'overlay'
  | 'saturation'
  | 'screen'
  | 'soft-light'

export type JsonShapeProperties = JsonBaseElementProperties & {
  backgroundColor?: JsonBackground
  type: 'circle' | 'rectangle'
  border?: JsonBorder
  cropData?: JsonCropData
  blendMode?: JsonBlendMode
}

export type JsonShape = {
  type: 'layer'
  layerType: 'shape'
  properties: JsonShapeProperties
}

export type JsonYoutubeProperties = JsonBaseElementProperties & {
  youtubeURL: string
  autoplay: boolean
  loop: boolean
  adjustColor?: JsonAdjustColor
  displayYoutubeLogo: boolean
  soundOnMouseOver: boolean
  volume: number
  startAt: boolean
  startAtTime: string
  hideControls: boolean
}

export type JsonYoutube = {
  type: 'layer'
  layerType: 'youtube'
  properties: JsonYoutubeProperties
}

export type JsonUseIframe = 'auto' | 'yes' | 'no'

export type JsonEmbedProperties = JsonBaseElementProperties & {
  backgroundColor?: JsonBackground
  code: string
  useBorder: boolean
  borderColor?: string
  margin: number
  originalWidth?: number
  useIframe: JsonUseIframe
  scrollbar: boolean
  autoResize: boolean
  adjustColor?: JsonAdjustColor
}

export type JsonEmbed = {
  type: 'layer'
  layerType: 'embed'
  properties: JsonEmbedProperties
}

export type JsonAutoplay = 'enabled' | 'disabled'

export type JsonVideoProperties = JsonBaseElementProperties & {
  format: string
  url: string
  scaleMode: JsonScaleMode
  horizontalAlign: JsonAlign
  verticalAlign: JsonAlign
  adjustColor?: JsonAdjustColor
  contentScale: number
  originalWidth: number
  originalHeight: number
  cropData?: JsonCropData
  startTime: number
  endTime: number
  duration: number
  autoplay: JsonAutoplay
  controls: boolean
  loop: boolean
  soundOnHover: boolean
  volume: number
  provider?: JsonStockMediaProvider
  maskImage?: string
  source?: string
}

export type JsonStockMediaProvider =
  | 'unsplash'
  | 'pixabay'
  | 'pexels'
  | 'giphy'
  | 'melodie'
  | 'audioProviders'
  | 'shutterstock'

export type JsonVideo = {
  type: 'layer'
  layerType: 'video'
  properties: JsonVideoProperties
}

export type JsonAudioProperties = JsonBaseElementProperties & {
  format: string
  url: string
  startTime: number
  endTime: number
  duration: number
  volume: number
  soundOnHover: boolean
}

export type JsonAudio = {
  type: 'layer'
  layerType: 'audio'
  properties: JsonAudioProperties
}

export type JsonClipartProperties = JsonBaseElementProperties & {
  // eslint-disable-next-line @typescript-eslint/no-explicit-any
  svgObject: any
}
// ...
export enum TextDecorationEnum {
  NONE = 'none',
  UNDERLINE = 'underline',
  STRIKETHROUGH = 'line-through',
}